Add files via upload
Day by day people shopping via E-commerce sites are burgeoning. Decision of placing
orders rely on product and service reviews provided by the customers. The importance of
the reviews has increased tremendously because they provide information about the quality
of product and service. This stimulates sellers to exploit these reviews to increase their sales
by deceiving the customers with false information. With the intention of increasing quality
service provided by E-sites and also make it easier for customers in selecting a product a
model was developed for detecting and preventing fake reviews.